Update HW4_Robotics_e.m
Dynamic Movement Primitives (DMPs) are a method of trajectory control / planning from Stefan Schaal’s lab. Complex movements have long been thought to be composed of sets of primitive action ‘building blocks’ executed in sequence and / or in parallel, and DMPs are a proposed mathematical formalization of these primitives. The difference between DMPs and previously proposed building blocks is that each DMP is a nonlinear dynamical system. The basic idea is that you take a dynamical system with well specified, stable behaviour and add another term that makes it follow some interesting trajectory as it goes about its business. A set of Differential Equations (Transformation System, Canonical System and Non- linear Function) realize a general way of generating point-to-point movements.

The weight factor W is computed by Imitation learning using linear regression on a demonstrated trajectory given by a teacher.